+++ server/plugins/config/forge.go
@@ -29,6 +29,39 @@ import (
 	"go.woodpecker-ci.org/woodpecker/v2/shared/constant"
 )
 
+type forgeFetcher struct {
+	timeout time.Duration
+}
+
+func NewForge(timeout time.Duration) Service {
+	return &forgeFetcher{
+		timeout: timeout,
+	}
+}
+
+func (f *forgeFetcher) Fetch(ctx context.Context, forge forge.Forge, user *model.User, repo *model.Repo, pipeline *model.Pipeline) (files []*types.FileMeta, err error) {
+	configFetcher := &forgeFetcherContext{
+		forge:    forge,
+		user:     user,
+		repo:     repo,
+		pipeline: pipeline,
+		timeout:  f.timeout,
+	}
+
+	// try to fetch 3 times
+	for i := 0; i < 3; i++ {
+		files, err = configFetcher.fetch(ctx, strings.TrimSpace(repo.Config))
+		if err != nil {
+			log.Trace().Err(err).Msgf("%d. try failed", i+1)
+		}
+		if errors.Is(err, context.DeadlineExceeded) {
+			continue
+		}
+	}
+
+	return
+}
+
 type forgeFetcherContext struct {
 	forge    forge.Forge
 	user     *model.User
